feat: add dynamic slots to gcds-table
