+++ packages/angular/src/lib/stencil-generated/index.ts
@@ -39,7 +39,6 @@ export const DIRECTIVES = [
   d.GcdsSrOnly,
   d.GcdsStepper,
   d.GcdsTable,
-  d.GcdsTableSlots,
   d.GcdsText,
   d.GcdsTextarea,
   d.GcdsTopNav,
