+++ packages/vue/tests/app/src/views/Table.vue
@@ -1,6 +1,6 @@
 <script setup>
 import { ref, onMounted, h } from 'vue'
-import { GcdsTableWithSlots } from '@gcds-core/components-vue'
+import { GcdsTable } from '@gcds-core/components-vue'
 
 // the first argument must match the ref value in the template
 const input = ref(null)
@@ -39,84 +39,6 @@ async function getFirst151Pokemon() {
 }
 
 onMounted(async () => {
-    // Populate the table columns
-    table.value.columns = [
-      {
-        field: 'number',
-        header: 'Pokédex',
-        align: 'end',
-        rowHeader: true,
-      },
-      {
-        field: 'name',
-        header: 'Name',
-      },
-      {
-        field: 'sprite',
-        header: 'Sprite',
-        align: 'center',
-        sort: false,
-        renderCell: (value, row) => {
-          const img = document.createElement('img');
-          img.src = `https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/${row.number}.png`;
-          img.alt = row.name;
-          img.width = 150;
-          img.height = 150;
-          return img;
-        },
-      },
-      { field: 'height', header: 'Height', align: 'end' },
-      { field: 'weight', header: 'Weight', align: 'end' },
-      {
-        field: 'base_experience',
-        header: 'Base experience',
-        sort: false,
-        align: 'end',
-      },
-      {
-        field: 'actions',
-        header: 'Actions',
-        align: 'center',
-        sort: false,
-        // Won't work in Vue
-        // renderCell: (_, row) => {
-        //   return h(
-        //     'gcds-button',
-        //     {
-        //       buttonRole: 'secondary',
-        //       onClick: () => {
-        //         if (row.number === 7) {
-        //           alert(`You clicked on ${row.name}! The best Pokémon!`);
-        //         } else {
-        //           alert(`This is ${row.name}, not Squirtle!`);
-        //         }
-        //       },
-        //     },
-        //     'Alert name',
-        //   );
-        // }
-        renderCell: (value, row) => {
-          const button = document.createElement('gcds-button');
-          button.setAttribute('button-role', 'secondary');
-          button.textContent = 'Alert name';
-          button.addEventListener('click', () => {
-            if (row.number === 7) {
-              alert(`You clicked on ${row.name}! The best Pokémon!`);
-            } else {
-              alert(`This is ${row.name}, not Squirtle!`);
-            }
-          });
-          return button;
-        },
-      },
-    ];
-
-
-  // Populate the table rows with results from API
-  getFirst151Pokemon().then(pokemon => {
-    table.value.data = pokemon;
-  });
-
   slotData.value = await getFirst151Pokemon();
 })
 </script>
@@ -134,9 +56,7 @@ onMounted(async () => {
       <gcds-heading tag="h2">Gcds-table</gcds-heading>
       <p>Testing whether the table still functions properly in Vue.</p>
 
-      <gcds-heading tag="h4">With slots</gcds-heading>
-
-      <GcdsTableWithSlots
+      <GcdsTable
         ref="tableSlot"
         id="my-table-slots"
         :pagination="true"
@@ -179,7 +99,7 @@ onMounted(async () => {
         ]"
       >
         <div slot="caption">
-          <gcds-heading tag="h4">Pokémon - Slots</gcds-heading>
+          <gcds-heading tag="h3">Pokémon</gcds-heading>
           Table of the best Pokémon (first generation).
         </div>
         <template #cell:sprite="{ row }">
@@ -201,22 +121,6 @@ onMounted(async () => {
             Console log row data
           </gcds-button>
         </template>
-      </GcdsTableWithSlots>
-
-      <gcds-heading tag="h3">With renderCell</gcds-heading>
-
-      <gcds-table
-        id="my-table"
-        ref="table"
-        sort="true"
-        pagination="true"
-        pagination-current-page="1"
-        filter="true"
-      >
-        <div slot="caption">
-          <gcds-heading tag="h4">Pokémon</gcds-heading>
-          Table of the best Pokémon (first generation).
-        </div>
-      </gcds-table>
+      </GcdsTable>
     </section>
 </template>
