+++ packages/web/src/components.d.ts
@@ -11,14 +11,14 @@ import { SpacingValues } from "./utils/types/spacing";
 import { ContentValues, GridGapValues } from "./components/gcds-grid/gcds-grid";
 import { SuggestionOption } from "./components/gcds-input/suggestion-option";
 import { RadioObject } from "./components/gcds-radios/radio";
-import { GcdsTableStateChange, TableColumn, TableColumnSlots } from "./components/gcds-table/utils/table-helpers";
+import { GcdsTableStateChange, TableColumn } from "./components/gcds-table/utils/table-helpers";
 export { CheckboxObject } from "./components/gcds-checkboxes/checkbox";
 export { Validator, ValidatorEntry } from "./validators";
 export { SpacingValues } from "./utils/types/spacing";
 export { ContentValues, GridGapValues } from "./components/gcds-grid/gcds-grid";
 export { SuggestionOption } from "./components/gcds-input/suggestion-option";
 export { RadioObject } from "./components/gcds-radios/radio";
-export { GcdsTableStateChange, TableColumn, TableColumnSlots } from "./components/gcds-table/utils/table-helpers";
+export { GcdsTableStateChange, TableColumn } from "./components/gcds-table/utils/table-helpers";
 export namespace Components {
     /**
      * Alert displays an alert message with an optional heading, icon, and close button.
@@ -1386,53 +1386,6 @@ export namespace Components {
           * @default ''
          */
         "filterValue": string;
-        /**
-          * Enable pagination
-          * @default false
-         */
-        "pagination": boolean;
-        /**
-          * Current page index
-          * @default 1
-         */
-        "paginationCurrentPage": number;
-        /**
-          * Number of rows per page
-          * @default 10
-         */
-        "paginationSize": number;
-        /**
-          * Available page-size options. Use 0 to represent "All rows".
-          * @default [     10, 25, 50, 0,   ]
-         */
-        "paginationSizeOptions": string | number[];
-        /**
-          * Enable global column sorting (can be overridden per column)
-          * @default false
-         */
-        "sort": boolean;
-    }
-    interface GcdsTableSlots {
-        /**
-          * Column definitions
-          * @default []
-         */
-        "columns": string | TableColumnSlots[];
-        /**
-          * Row data
-          * @default []
-         */
-        "data": string | object[];
-        /**
-          * Enable global filter
-          * @default false
-         */
-        "filter": boolean;
-        /**
-          * Current filter string
-          * @default ''
-         */
-        "filterValue": string;
         "getVisibleRows": () => Promise<{ rowId: string; rowIndex: number; original: Record<string, unknown>; }[]>;
         /**
           * Enable pagination
@@ -1724,9 +1677,9 @@ export interface GcdsSelectCustomEvent<T> extends CustomEvent<T> {
     detail: T;
     target: HTMLGcdsSelectElement;
 }
-export interface GcdsTableSlotsCustomEvent<T> extends CustomEvent<T> {
+export interface GcdsTableCustomEvent<T> extends CustomEvent<T> {
     detail: T;
-    target: HTMLGcdsTableSlotsElement;
+    target: HTMLGcdsTableElement;
 }
 export interface GcdsTextareaCustomEvent<T> extends CustomEvent<T> {
     detail: T;
@@ -2338,28 +2291,22 @@ declare global {
         prototype: HTMLGcdsStepperElement;
         new (): HTMLGcdsStepperElement;
     };
-    interface HTMLGcdsTableElement extends Components.GcdsTable, HTMLStencilElement {
-    }
-    var HTMLGcdsTableElement: {
-        prototype: HTMLGcdsTableElement;
-        new (): HTMLGcdsTableElement;
-    };
-    interface HTMLGcdsTableSlotsElementEventMap {
+    interface HTMLGcdsTableElementEventMap {
         "gcdsTableStateChange": GcdsTableStateChange;
     }
-    interface HTMLGcdsTableSlotsElement extends Components.GcdsTableSlots, HTMLStencilElement {
-        addEventListener<K extends keyof HTMLGcdsTableSlotsElementEventMap>(type: K, listener: (this: HTMLGcdsTableSlotsElement, ev: GcdsTableSlotsCustomEvent<HTMLGcdsTableSlotsElementEventMap[K]>) => any, options?: boolean | AddEventListenerOptions): void;
+    interface HTMLGcdsTableElement extends Components.GcdsTable, HTMLStencilElement {
+        addEventListener<K extends keyof HTMLGcdsTableElementEventMap>(type: K, listener: (this: HTMLGcdsTableElement, ev: GcdsTableCustomEvent<HTMLGcdsTableElementEventMap[K]>) => any, options?: boolean | AddEventListenerOptions): void;
         addEventListener<K extends keyof DocumentEventMap>(type: K, listener: (this: Document, ev: DocumentEventMap[K]) => any, options?: boolean | AddEventListenerOptions): void;
         addEventListener<K extends keyof HTMLElementEventMap>(type: K, listener: (this: HTMLElement, ev: HTMLElementEventMap[K]) => any, options?: boolean | AddEventListenerOptions): void;
         addEventListener(type: string, listener: EventListenerOrEventListenerObject, options?: boolean | AddEventListenerOptions): void;
-        removeEventListener<K extends keyof HTMLGcdsTableSlotsElementEventMap>(type: K, listener: (this: HTMLGcdsTableSlotsElement, ev: GcdsTableSlotsCustomEvent<HTMLGcdsTableSlotsElementEventMap[K]>) => any, options?: boolean | EventListenerOptions): void;
+        removeEventListener<K extends keyof HTMLGcdsTableElementEventMap>(type: K, listener: (this: HTMLGcdsTableElement, ev: GcdsTableCustomEvent<HTMLGcdsTableElementEventMap[K]>) => any, options?: boolean | EventListenerOptions): void;
         removeEventListener<K extends keyof DocumentEventMap>(type: K, listener: (this: Document, ev: DocumentEventMap[K]) => any, options?: boolean | EventListenerOptions): void;
         removeEventListener<K extends keyof HTMLElementEventMap>(type: K, listener: (this: HTMLElement, ev: HTMLElementEventMap[K]) => any, options?: boolean | EventListenerOptions): void;
         removeEventListener(type: string, listener: EventListenerOrEventListenerObject, options?: boolean | EventListenerOptions): void;
     }
-    var HTMLGcdsTableSlotsElement: {
-        prototype: HTMLGcdsTableSlotsElement;
-        new (): HTMLGcdsTableSlotsElement;
+    var HTMLGcdsTableElement: {
+        prototype: HTMLGcdsTableElement;
+        new (): HTMLGcdsTableElement;
     };
     /**
      * Text is a styled and formatted paragraph that displays written content in an accessible way and matches Canada.ca typography styles.
@@ -2451,7 +2398,6 @@ declare global {
         "gcds-sr-only": HTMLGcdsSrOnlyElement;
         "gcds-stepper": HTMLGcdsStepperElement;
         "gcds-table": HTMLGcdsTableElement;
-        "gcds-table-slots": HTMLGcdsTableSlotsElement;
         "gcds-text": HTMLGcdsTextElement;
         "gcds-textarea": HTMLGcdsTextareaElement;
         "gcds-top-nav": HTMLGcdsTopNavElement;
@@ -4058,54 +4004,7 @@ declare namespace LocalJSX {
           * @default ''
          */
         "filterValue"?: string;
-        /**
-          * Enable pagination
-          * @default false
-         */
-        "pagination"?: boolean;
-        /**
-          * Current page index
-          * @default 1
-         */
-        "paginationCurrentPage"?: number;
-        /**
-          * Number of rows per page
-          * @default 10
-         */
-        "paginationSize"?: number;
-        /**
-          * Available page-size options. Use 0 to represent "All rows".
-          * @default [     10, 25, 50, 0,   ]
-         */
-        "paginationSizeOptions"?: string | number[];
-        /**
-          * Enable global column sorting (can be overridden per column)
-          * @default false
-         */
-        "sort"?: boolean;
-    }
-    interface GcdsTableSlots {
-        /**
-          * Column definitions
-          * @default []
-         */
-        "columns"?: string | TableColumnSlots[];
-        /**
-          * Row data
-          * @default []
-         */
-        "data"?: string | object[];
-        /**
-          * Enable global filter
-          * @default false
-         */
-        "filter"?: boolean;
-        /**
-          * Current filter string
-          * @default ''
-         */
-        "filterValue"?: string;
-        "onGcdsTableStateChange"?: (event: GcdsTableSlotsCustomEvent<GcdsTableStateChange>) => void;
+        "onGcdsTableStateChange"?: (event: GcdsTableCustomEvent<GcdsTableStateChange>) => void;
         /**
           * Enable pagination
           * @default false
@@ -4348,7 +4247,6 @@ declare namespace LocalJSX {
         "gcds-sr-only": GcdsSrOnly;
         "gcds-stepper": GcdsStepper;
         "gcds-table": GcdsTable;
-        "gcds-table-slots": GcdsTableSlots;
         "gcds-text": GcdsText;
         "gcds-textarea": GcdsTextarea;
         "gcds-top-nav": GcdsTopNav;
@@ -4504,7 +4402,6 @@ declare module "@stencil/core" {
              */
             "gcds-stepper": LocalJSX.GcdsStepper & JSXBase.HTMLAttributes<HTMLGcdsStepperElement>;
             "gcds-table": LocalJSX.GcdsTable & JSXBase.HTMLAttributes<HTMLGcdsTableElement>;
-            "gcds-table-slots": LocalJSX.GcdsTableSlots & JSXBase.HTMLAttributes<HTMLGcdsTableSlotsElement>;
             /**
              * Text is a styled and formatted paragraph that displays written content in an accessible way and matches Canada.ca typography styles.
              */
