+++ packages/web/src/components/gcds-heading/readme.md
@@ -35,7 +35,6 @@ A heading is a title that establishes levels of hierarchy to organize page conte
  - [gcds-notice](../gcds-notice)
  - [gcds-stepper](../gcds-stepper)
  - [gcds-table](../gcds-table)
- - [gcds-table-slots](../gcds-table-slots)
 
 ### Graph
 ```mermaid
@@ -44,7 +43,6 @@ graph TD;
   gcds-notice --> gcds-heading
   gcds-stepper --> gcds-heading
   gcds-table --> gcds-heading
-  gcds-table-slots --> gcds-heading
   style gcds-heading fill:#26374a,stroke:#333,stroke-width:4px
 ```
