+++ packages/web/src/components/gcds-select/readme.md
@@ -96,7 +96,6 @@ Type: `Promise<void>`
 
  - [gcds-date-input](../gcds-date-input)
  - [gcds-table](../gcds-table)
- - [gcds-table-slots](../gcds-table-slots)
 
 ### Depends on
 
@@ -115,7 +114,6 @@ graph TD;
   gcds-error-message --> gcds-icon
   gcds-date-input --> gcds-select
   gcds-table --> gcds-select
-  gcds-table-slots --> gcds-select
   style gcds-select fill:#26374a,stroke:#333,stroke-width:4px
 ```
