+++ packages/web/src/components/gcds-sr-only/readme.md
@@ -38,7 +38,6 @@ The screenreader-only component is text information only accessible with assisti
  - [gcds-signature](../gcds-signature)
  - [gcds-stepper](../gcds-stepper)
  - [gcds-table](../gcds-table)
- - [gcds-table-slots](../gcds-table-slots)
  - [gcds-textarea](../gcds-textarea)
  - [gcds-topic-menu](../gcds-topic-menu)
 
@@ -56,7 +55,6 @@ graph TD;
   gcds-signature --> gcds-sr-only
   gcds-stepper --> gcds-sr-only
   gcds-table --> gcds-sr-only
-  gcds-table-slots --> gcds-sr-only
   gcds-textarea --> gcds-sr-only
   gcds-topic-menu --> gcds-sr-only
   style gcds-sr-only fill:#26374a,stroke:#333,stroke-width:4px
