+++ packages/web/src/components/gcds-table/test/gcds-table.spec.tsx
@@ -95,6 +95,7 @@ describe('gcds-table', () => {
               </tbody>
             </table>
           </section>
+          <slot></slot>
         </mock:shadow-root>
       </gcds-table>
     `);
@@ -110,9 +111,9 @@ describe('gcds-table', () => {
   it('renders caption slot content', async () => {
     const page = await setup(
       '<gcds-table><span slot="caption">' +
-        '<h2>Pokémon</h2>' +
-        'Table of the best Pokémon (first generation).' +
-        '</span></gcds-table>',
+      '<h2>Pokémon</h2>' +
+      'Table of the best Pokémon (first generation).' +
+      '</span></gcds-table>',
     );
 
     expect(
@@ -200,7 +201,7 @@ describe('gcds-table', () => {
     expect(
       page.root?.shadowRoot?.querySelector('tbody tr td[data-column="Name"]'),
     ).toEqualHtml(`
-    <td class="gcds-table__td" data-column="Name">
+    <td class="gcds-table__td" data-cell="name-0" data-column="Name">
       &lt;script&gt;alert('XSS attack');&lt;/script&gt;
     </td>
     `);
@@ -277,26 +278,6 @@ describe('gcds-table', () => {
     // expect height to be td and not th
     expect(heightCell?.tagName).toBe('TD');
     expect(heightRowHeader).toBeNull();
-  });
-
-  it('invokes renderCell callback with value and row', async () => {
-    const renderCell = jest.fn((value: unknown) => String(value));
-    const page = await setup();
-    const el = page.root as HTMLGcdsTableElement;
-
-    el.columns = [
-      { field: 'number', header: 'Pokédex' },
-      { field: 'name', header: 'Name', renderCell },
-    ] as any;
-    el.data = baseData as any;
-    await page.waitForChanges();
-
-    expect(renderCell).toHaveBeenCalled();
-
-    // TODO: Expect to render x how many rows there are (3 in this case)
-    // However, right now it's being called 9 times, so this needs to be revisited
-    // expect(renderCell).toHaveBeenCalledTimes(3);
-    expect(renderCell).toHaveBeenNthCalledWith(1, 'Squirtle', baseData[0]);
   });
 
   it('renders empty state text when there is no data', async () => {
@@ -787,5 +768,128 @@ describe('gcds-table', () => {
     const instance = page.rootInstance as any;
     expect(instance.table.getPageCount()).toBe(2);
     expect(instance.paginationState.pageIndex).toBe(0);
+  });
+
+  it('renders slotted element', async () => {
+    const page = await setup();
+    const el = page.root as HTMLGcdsTableElement;
+
+    // Extend base columns
+    const extendedColumns = [
+      ...baseColumns,
+      {
+        field: 'sprite',
+        header: 'Sprite',
+        slotted: true,
+      },
+    ];
+
+    const templateElement = document.createElement('template');
+    templateElement.setAttribute('slot', 'cell:sprite');
+
+    const img = document.createElement('img');
+    img.setAttribute('src', 'https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/7.png');
+    img.setAttribute('alt', 'Squirtle sprite');
+    templateElement.content.appendChild(img);
+
+    el.appendChild(templateElement);
+
+    el.columns = extendedColumns as any;
+    el.data = baseData as any;
+    await page.waitForChanges();
+
+    const headers = page.root?.shadowRoot?.querySelectorAll('thead th') || [];
+    const rows = page.root?.shadowRoot?.querySelectorAll('tbody tr') || [];
+    const firstRowCells = rows[0]?.querySelectorAll('td') || [];
+
+    expect(headers).toHaveLength(6);
+    expect(headers[0].textContent).toContain('Pokédex');
+    expect(rows).toHaveLength(3);
+
+    expect(firstRowCells[0]?.textContent?.trim()).toBe('7');
+    expect(firstRowCells[1]?.textContent?.trim()).toBe('Squirtle');
+    expect(firstRowCells[2]?.textContent?.trim()).toBe('5');
+    expect(firstRowCells[3]?.textContent?.trim()).toBe('90');
+    expect(firstRowCells[4]?.textContent?.trim()).toBe('63');
+    expect(firstRowCells[5]).toEqualHtml(`<td class="gcds-table__td" data-cell="sprite-0" data-column="Sprite">
+        <div>
+          <img data-column-key="sprite" data-row-id="0" src="https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/7.png" alt="Squirtle sprite">
+        </div>
+      </td>`);
+  });
+
+  it('renders slotted element with data-bind-* attributes', async () => {
+    const page = await setup();
+    const el = page.root as HTMLGcdsTableElement;
+
+    // Extend base columns
+    const extendedColumns = [
+      ...baseColumns,
+      {
+        field: 'sprite',
+        header: 'Sprite',
+        slotted: true,
+      },
+    ];
+
+    // Create template element with data-bind attributes that reference row data fields
+    const templateElement = document.createElement('template');
+    templateElement.setAttribute('slot', 'cell:sprite');
+
+    const img = document.createElement('img');
+    img.setAttribute('data-bind-template-src', 'https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/{number}.png');
+    img.setAttribute('data-bind-alt', 'name');
+    templateElement.content.appendChild(img);
+
+    el.appendChild(templateElement);
+
+    el.columns = extendedColumns as any;
+    el.data = baseData as any;
+    await page.waitForChanges();
+
+    const headers = page.root?.shadowRoot?.querySelectorAll('thead th') || [];
+    const rows = page.root?.shadowRoot?.querySelectorAll('tbody tr') || [];
+
+    expect(headers).toHaveLength(6);
+    expect(rows).toHaveLength(3);
+
+    expect(rows[0]?.querySelectorAll('td')[5]).toEqualHtml(`<td class="gcds-table__td" data-cell="sprite-0" data-column="Sprite">
+        <div>
+          <img
+            data-column-key="sprite"
+            data-bind-alt="name"
+            data-bind-template-src="https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/{number}.png"
+            data-row-id="0"
+            src="https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/7.png"
+            alt="Squirtle"
+          />
+        </div>
+      </td>`);
+
+    expect(rows[1]?.querySelectorAll('td')[5]).toEqualHtml(`<td class="gcds-table__td" data-cell="sprite-1" data-column="Sprite">
+        <div>
+          <img
+            data-column-key="sprite"
+            data-bind-alt="name"
+            data-bind-template-src="https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/{number}.png"
+            data-row-id="1"
+            src="https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/8.png"
+            alt="Wartortle"
+          />
+        </div>
+      </td>`);
+
+    expect(rows[2]?.querySelectorAll('td')[5]).toEqualHtml(`<td class="gcds-table__td" data-cell="sprite-2" data-column="Sprite">
+        <div>
+          <img
+            data-column-key="sprite"
+            data-bind-alt="name"
+            data-bind-template-src="https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/{number}.png"
+            data-row-id="2"
+            src="https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/9.png"
+            alt="Blastoise"
+          />
+        </div>
+      </td>`);
   });
 });
